+++ packages/ridb-level/release.config.js
@@ -1,51 +1,14 @@
+import  {branches, plugins} from '../../release.base.js'
 export default {
-  tagFormat: '@trust0/ridb@${version}',
-    "branches": [
+  tagFormat: '@trust0/ridb-level@${version}',
+  branches: branches,
+  plugins: [
+    [
+      "@semantic-release/exec",
       {
-        "name": "main"
-      },
-      {
-        "name": "develop",
-        "prerelease": "rc",
-        "channel": "rc"
-      },
-      "v+([0-9])?(.{+([0-9]),x}).x"
+        "prepareCmd": "sh ../../update.sh"
+      }
     ],
-    "plugins": [
-      "@semantic-release/changelog",
-      [
-        "@semantic-release/commit-analyzer",
-        {
-          "preset": "conventionalcommits"
-        }
-      ],
-      [
-        "@semantic-release/release-notes-generator",
-        {
-          "preset": "conventionalcommits"
-        }
-      ],
-      "@semantic-release/github",
-      [
-        "@semantic-release/exec",
-        {
-          "prepareCmd": "npx npm-check-updates -u @trust0/ridb -t newest && npx npm-check-updates -u @trust0/ridb-core -t newest && npm i"
-        }
-      ],
-      "@semantic-release/npm",
-      [
-        "@semantic-release/git",
-        {
-          "assets": [
-            "../../package-lock.json",
-            "../../package.json",
-            "package-lock.json",
-            "package.json",
-            "CHANGELOG.md",
-            "docs/**/*"
-          ],
-          "message": "chore(release): ${nextRelease.version} [skip ci]\n\n${nextRelease.notes}"
-        }
-      ]
-    ],
-  }
+    ...plugins,
+  ],
+}